refactor: rename features to contract and add new interfaces
