+++ contract/nameable.go
@@ -1,0 +1,5 @@
+package contract
+
+type Nameable interface {
+	Name() string
+}
