+++ contract/retriable.go
@@ -1,0 +1,5 @@
+package contract
+
+type Retriable interface {
+	Retries() int
+}
