--- features/async_test.go
@@ -1,13 +1,0 @@
-package features
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestAsync(t *testing.T) {
-	f := AsyncFeature{}
-
-	assert.True(t, f.Async())
-}
