--- features/named_test.go
@@ -1,13 +1,0 @@
-package features
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestNamedFeature(t *testing.T) {
-	f := NewNamedFeature("test")
-
-	assert.Equal(t, "test", f.Name())
-}
