--- features/slug.go
@@ -1,17 +1,0 @@
-package features
-
-type Slug interface {
-	Slug() string
-}
-
-type SlugFeature struct {
-	slug string
-}
-
-func NewSlugFeature(slug string) *SlugFeature {
-	return &SlugFeature{slug: slug}
-}
-
-func (n *SlugFeature) Slug() string {
-	return n.slug
-}
